+++ src/homework1/HelloWorld.java
@@ -1,4 +1,4 @@
-package HomeWork1;
+package homework1;
 
 public class HelloWorld {
