+++ src/homework1/Task2.java
@@ -1,4 +1,4 @@
-package HomeWork1;
+package homework1;
 
 public class Task2 {
     public static void main(String[] args) {
