+++ src/homework1/Task3.java
@@ -1,4 +1,4 @@
-package HomeWork1;
+package homework1;
 /*
 Имеется 4500 секунд. Вывести в консоль содержащихся в этом количестве секунд:
  А) минут + секунд,
