+++ src/homework1/Task4.java
@@ -1,4 +1,4 @@
-package HomeWork1;
+package homework1;
 
 /*
 Определить число, полученное выписыванием в обратном порядке цифр любого 4-х значного натурального числа n.
