+++ src/homework1/Task5.java
@@ -1,4 +1,4 @@
-package HomeWork1;
+package homework1;
 /*
 Дано любое натуральное 4-х значное число. Верно ли, что все цифры числа различны?
  */
