+++ src/homework2/Task10.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
 10) Посчитать факториал числа в границах от 10 до 15 (не используя рекурсию).
