+++ src/homework2/Task11.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
  11) Имеется целое число, определить является ли это число простым, т.е.
