+++ src/homework2/Task13.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
  13) Создать последовательность случайных чисел, найти и вывести наибольшее из них.
