+++ src/homework2/Task15.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
  15) Создать массив, заполнить его случайными элементами, распечатать, перевернуть,
