+++ src/homework2/Task17.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
  17) Создать двухмерный квадратный массив, и заполнить его «бабочкой», т.е таким образом:
