+++ src/homework2/Task6.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 /*
 Создайте число. Определите, является ли число трехзначным. Определите, является
 ли его последняя цифра семеркой. Определите, является ли число четным.
