+++ src/homework2/Task7.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
 7) Имеется прямоугольное отверстие размерами a и b, определить, можно ли
