+++ src/homework2/Task9.java
@@ -1,4 +1,4 @@
-package homeWork2;
+package homework2;
 
 /*
 9) Изменить пример с суммой чисел таким образом, чтобы рассчитывалась не сумма, а
